Use transparent backgroud for urxvt (i3 setup)
When using urxvt on i3, transparency can be enabled by
installing a compositor. However the chroma colorscheme
still has a solid gray background so transparency doesn't
work right off the bat (it works on yakuake and konsole
for some reason). The workaround is to override the
theme's bg color with NONE as set here. I also changed the
color of the CursorLine to be more visible in this setting.